+++ brewtils/rest/publish_client.py
@@ -78,7 +78,7 @@ class PublishClient(object):
 
         Args:
             _topic (str): The topic to publish to, default is Plugin level topic
-                (Namespace.System.Version.Instance)
+                {Namespace}.{System}.{Version}.{Instance}
             _regex_only (bool): If the request will be resolved against only annotated topics
             from the @subscribe command
             _propagate (bool): If the request will be pushed up to the parent to be resolved.
@@ -92,9 +92,10 @@ class PublishClient(object):
                 brewtils.plugin.CONFIG.name
                 and brewtils.plugin.CONFIG.version
                 and brewtils.plugin.CONFIG.instance_name
+                and brewtils.plugin.CONFIG.namespace
             ):
                 _topic = "{0}.{1}.{2}.{3}".format(
-                    brewtils.plugin.CONFIG.namespace or "",
+                    brewtils.plugin.CONFIG.namespace,
                     brewtils.plugin.CONFIG.name,
                     brewtils.plugin.CONFIG.version,
                     brewtils.plugin.CONFIG.instance_name,