+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Brewtils Changelog
 TBD
 
 - Added new KWARG input to @command for tag/tags. This can be utilized for filtering commands.
-- Adding defaul topic for PublishClient to Plugins {Namespace}.{System}.{Version}.{Instance}
+- Adding default topic for PublishClient to Plugins {Namespace}.{System}.{Version}.{Instance}
 - Removed Python 12 support until we upgrade Marshmallow dependency to 3.15 or greater
 
 3.21.0